Material objects properties constructing logic updated.
MaterialProperties removed from objects properties extending.
Code analysed and fixed in some places.
Code copy-paste deleted from objects hierarchies super classes and other.
Some helping code wrote. Renaming.

+++ engine/main/objects/creation/scene/generators/reduction/CustomReduction.cpp
@@ -15,7 +15,7 @@ Creation::RD::loadAlgorithms(const IniUtil::Analyzer::IniBlock &data, Creation::
 std::shared_ptr<sf::Music>
 Creation::RD::loadMusic(const IniUtil::Analyzer::IniBlock &data, Creation::Pools::SourcePool &pool) {
     std::shared_ptr<sf::Music> music = pool.getMusic(data.at("AUDIO"));
-    music->setVolume(std::atof(data.at("AUDIO_VOLUME").c_str()));
+    music->setVolume(std::stof(data.at("AUDIO_VOLUME").c_str()));
     music->setLoop(static_cast<bool>(std::atoi(data.at("AUDIO_LOOP").c_str())));
     return music;
 }
